Fix Plugin memory leak causing extra launches
Every time a refresh occurred a new instance of a Plugin would be
created, with old instances being set to nil. The intention must have
been for ARC to clean it up, but somewhere an extra reference was
hodling on. The strong properties in ExecutablePlugin.h were suspect, as
well as the behavior of the repeating NSTimers in threads that
indefinietly spun up new dispatch_async calls.

I didn't try to find how to make that all work. Instead, I stopped the
PluginManager from creating new plugins every time refresh is called by
the user, which seemed needless. Now, plugins are only ever created
once, and user refresh requests just call the refresh method of each
plugin known plugin.

One side effect of this is that BitBar will need to be restarted in
order for it to pick up new plugins, but it simplifies the logic a lot.
A potential work around for that might be to just listen for file system
changes on the plugin directory and load the ones that are added.