Quick fix for memory leak causing extra runs
This addresses https://github.com/matryer/bitbar/issues/195

This is a slightly hacky fix for what looks like a memory leak issue
where plugins keep refreshing bitbar because they never get deallcoated,
regardless of setting all known references to them to `nil`.

This change will cease any further scheduling jobs in plugins if that
plugin is not in the list of plugins known to the PluginManager.