Potential fix for #88: Read from a config.json file
I wanted a way to separate configuration from code, so this is an
attempt to do so.

Reads configuration from a `config.json` file in the bitbar plugin
directory, and loads it into the environment of the script so that it's
available to plugins.

This is probably the most lines of Objective-C I've ever written in one
sitting. Feel free to critique.